Move all cache related tests in "right" place
- ServiceHealthCacheTests is moved into cache package.
- KVCache tests are moved in KVCacheTest
- HealthCheckCache tests are moved in HealthCheckCacheTest
- ServiceHealthCache tests are moved in ServiceHealthCacheTest
